working glfw example
